requirements.txt
Flask-RESTful<0.4
docker-compose==1.7.0
GitPython<3
